requirements.txt
flask==3.1.1
gunicorn==23.0.0
python-dotenv==1.1.0
pypdf==5.6.1

# LangChain core + integrations
langchain==0.3.26
langchain-community==0.3.26
langchain-pinecone==0.2.8
langchain-groq==0.1.3   # (check latest version available)

# Hugging Face API client (for embeddings)
huggingface-hub==0.24.6
